backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.3
feedparser>=6.0.11
crewai>=0.120.1
